+++ llama_index/chat_engine/types.py
@@ -125,6 +125,10 @@ class StreamingAgentChatResponse:
                 # Queue is empty, but we're not done yet
                 continue
 
+    def print_response_stream(self) -> None:
+        for token in self.response_gen:
+            print(token, end="")
+
 
 class BaseChatEngine(ABC):
     """Base Chat Engine."""
@@ -194,6 +198,13 @@ class ChatMode(str, Enum):
     
     First generate a standalone question from conversation context and last message,
     then query the query engine for a response.
+    """
+
+    CONTEXT = "context"
+    """Corresponds to `ContextChatEngine`.
+    
+    First retrieve text from the index using the user's message, then use the context
+    in the system prompt to generate a response.
     """
 
     REACT = "react"