+++ llama_index/indices/base_retriever.py
@@ -1,9 +1,9 @@
 from abc import ABC, abstractmethod
 from typing import List, Optional
 
-from llama_index.schema import NodeWithScore
 from llama_index.indices.query.schema import QueryBundle, QueryType
 from llama_index.indices.service_context import ServiceContext
+from llama_index.schema import NodeWithScore
 
 
 class BaseRetriever(ABC):
@@ -21,6 +21,12 @@ class BaseRetriever(ABC):
             str_or_query_bundle = QueryBundle(str_or_query_bundle)
         return self._retrieve(str_or_query_bundle)
 
+    async def aretrieve(self, str_or_query_bundle: QueryType) -> List[NodeWithScore]:
+        if isinstance(str_or_query_bundle, str):
+            str_or_query_bundle = QueryBundle(str_or_query_bundle)
+        nodes = await self._aretrieve(str_or_query_bundle)
+        return nodes
+
     @abstractmethod
     def _retrieve(self, query_bundle: QueryBundle) -> List[NodeWithScore]:
         """Retrieve nodes given query.
@@ -29,6 +35,16 @@ class BaseRetriever(ABC):
 
         """
         pass
+
+    # TODO: make this abstract
+    # @abstractmethod
+    async def _aretrieve(self, query_bundle: QueryBundle) -> List[NodeWithScore]:
+        """Asyncronously retrieve nodes given query.
+
+        Implemented by the user.
+
+        """
+        return []
 
     def get_service_context(self) -> Optional[ServiceContext]:
         """Attempts to resolve a service context.