+++ llama_index/langchain_helpers/text_splitter.py
@@ -1,12 +1,11 @@
 """Text splitter implementations."""
 from dataclasses import dataclass
-from typing import Callable, List, Optional
+from typing import Any, Callable, Dict, List, Optional
 
-from llama_index.constants import DEFAULT_CHUNK_OVERLAP, DEFAULT_CHUNK_SIZE
 from llama_index.bridge.langchain import TextSplitter
-
 from llama_index.callbacks.base import CallbackManager
 from llama_index.callbacks.schema import CBEventType, EventPayload
+from llama_index.constants import DEFAULT_CHUNK_OVERLAP, DEFAULT_CHUNK_SIZE
 from llama_index.utils import globals_helper
 
 
@@ -21,6 +20,7 @@ class TextSplit:
 
     text_chunk: str
     num_char_overlap: Optional[int] = None
+    metadata: Optional[Dict[str, Any]] = None
 
 
 class TokenTextSplitter(TextSplitter):