+++ llama_index/program/openai_program.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 
 from llama_index.llms.base import LLM, ChatMessage, MessageRole
 from llama_index.llms.openai import OpenAI
-from llama_index.llms.openai_utils import is_function_calling_model, to_openai_function
+from llama_index.llms.openai_utils import to_openai_function
 from llama_index.program.llm_prompt_program import BaseLLMFunctionProgram
 from llama_index.prompts.base import Prompt
 from llama_index.types import Model
@@ -18,7 +18,7 @@ def _default_function_call(output_cls: Type[BaseModel]) -> Dict[str, Any]:
     }
 
 
-class OpenAIPydanticProgram(BaseLLMFunctionProgram[OpenAI]):
+class OpenAIPydanticProgram(BaseLLMFunctionProgram[LLM]):
     """
     An OpenAI-based function that returns a pydantic model.
 
@@ -28,7 +28,7 @@ class OpenAIPydanticProgram(BaseLLMFunctionProgram[OpenAI]):
     def __init__(
         self,
         output_cls: Type[Model],
-        llm: OpenAI,
+        llm: LLM,
         prompt: Prompt,
         function_call: Union[str, Dict[str, Any]],
         verbose: bool = False,
@@ -51,12 +51,11 @@ class OpenAIPydanticProgram(BaseLLMFunctionProgram[OpenAI]):
         **kwargs: Any,
     ) -> "OpenAIPydanticProgram":
         llm = llm or OpenAI(model="gpt-3.5-turbo-0613")
-        if not isinstance(llm, OpenAI):
-            raise ValueError("llm must be a OpenAI instance")
 
-        if not is_function_calling_model(llm.model):
+        if not llm.metadata.is_function_calling_model:
             raise ValueError(
-                f"Model name {llm.model} does not support function calling API. "
+                f"Model name {llm.metadata.model_name} does not support "
+                "function calling API. "
             )
 
         prompt = Prompt(prompt_template_str)
@@ -100,7 +99,10 @@ class OpenAIPydanticProgram(BaseLLMFunctionProgram[OpenAI]):
             arguments_str = function_call["arguments"]
             print(f"Function call: {name} with args: {arguments_str}")
 
-        output = self.output_cls.parse_raw(function_call["arguments"])
+        if isinstance(function_call["arguments"], dict):
+            output = self.output_cls.parse_obj(function_call["arguments"])
+        else:
+            output = self.output_cls.parse_raw(function_call["arguments"])
         return output
 
     async def acall(
@@ -130,5 +132,8 @@ class OpenAIPydanticProgram(BaseLLMFunctionProgram[OpenAI]):
             arguments_str = function_call["arguments"]
             print(f"Function call: {name} with args: {arguments_str}")
 
-        output = self.output_cls.parse_raw(function_call["arguments"])
+        if isinstance(function_call["arguments"], dict):
+            output = self.output_cls.parse_obj(function_call["arguments"])
+        else:
+            output = self.output_cls.parse_raw(function_call["arguments"])
         return output